Completed additional assignment
As a user, I may want to add in a transaction that was missing, by passing it with a date in the past (eg. account.deposit(50, ’03/08/2020’) or whatever date format makes sense for your code).
This should in turn readjust my statement so that the transaction shows in the right place, and the balances are still correct. Does your code currently make this easy to add in? What would need change in your code to make this possible?

First of all this meant I had to change the account class to allow it to be passed a date. I then needed to order the history array based on the date of each transaction. This then highlighted that my balances are incorrect as they are correct for when the transaction was put into the array, not when the transaction was actually made. I removed the balance instance variable of the transaction class and instead calculated the balance when I printed the statement. Due to the formatting of the expected output, I had to calculate the balance accumulator in ascending date but display the statements in descending date.