Added american date feature
To be able to change the date formatting I needed to separate the formatting and the data. This meant storing the raw date inside the transaction and formatting in the print_statement method. This then allowed me to pass the option into the format_date method and decide which format to use.